UPL relicensing, stage 2: automated modifications.
This relicenses everything already containing a CDDL license
to use the short form of the UPL.

All Oracle copyright notices are reworded to the Oracle standard
form, and reshuffled above the license text.  (If you stick to it
precisely, it is never more than 79 chars long, which is a nice
touch since it means we don't need to word-wrap copyright lines.)

Long-obsolete ident pragmas, and useless exhortations to conform
to the license, are dropped, as was recently done for the kernel.

In addition, we fix up a few executables' execute bits, which
have been wrong forever (mostly in libdtrace and the testsuite).
(This happened at the same time because the script didn't preserve
execute bits, so I had to remark things as executable en masse.)