lexer: allow keywords after structure member dereference operators
We want to allow references to the numerous things named 'self' in the
Linux kernel, even though there is a D keyword of the same name.  There
is really no justification for prohibiting access to any keyword at all
in this position: nothing is syntactically valid after a -> or . other
than an identifier or comment either (the parser contains no rules for
other cases, and the code generator contains an assertion validating the
same thing).

The fix is simple: add a new start condition that indicates to the lexer
that we are inside a place where only an identifier (or comment) is
permitted, then conditionalize identifier matches on this condition and
reset it back to its original condition again.  We are lucky in that all
such places are in condition S0, so we don't need to specially track the
previous lexer condition (since the only thing that tracks such things
is already in use for comments.)

This is not terribly elegant, but it works, it's no uglier than the
things the lexer is doing anyway, and we are unlikely to need to
extend it in future, since the chance of extra things being permitted,
C++-style, after . or -> is remote.

Orabug: 26518086