tests: err.unloaded_var.x: cater for CTF archives
This checks for the presence of a .ko to tell whether CTF is present for
an unloaded module: if the module is not there, the CTF is assumed to
not be loadable and the test is XFAIL.

This condition is wrong for a CTF archive world, where the CTF is
accessible even if the module is not, and we are spuriously breaking now
that modules are xz-compressed so their name has changed.

Adjust accordingly.