dtrace: dtrace: sys/procfs.h wrapper does not include systems procfs.h
The <sys/procfs.h> wrapper beaks build on some architectures. The
problem is that it uses -Iinclude to interpose on the system wide
header. When it tries to include system wide header it ends up
including itself.

To fix this a new name has been used for the wrapper itself. Now
when the real procfs.h gets pulled in we have to drop some conflicting
defines in dtrace_types.h

Orabug: 27266725

Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>