libproc: note that processes are released before we unlock
This avoids a narrow race window where a Prelease() triggered
by a terminated process or shutting-down dtrace process-
control thread collides with the Pfree() in dt_ps_proc_destroy(),
leading to an assertion failure because P->released is still
false.

So ensure that everything Prelease() needs to do to the process
is done before the unlock happens.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Orabug: 26848964