Improve handling of symbols read from /proc/kallmodsyms
In dt_module.c, dt_modsym_update() attributed weak symbols to data.
So when an address corresponded to both a text and a weak symbol,
a text and data segment would overlap.  Instead, we should attribute
weak symbols to text, as was already done in sym_type_to_info().

The function also blacklisted symbols with "." in their names.
That was incorrect.

The function should also drop 'a' and 'A' symbols.  They tend to be
allocated per-CPU and the reported "addresses" are actually offsets
that we do not yet handle properly.  This change means dt_modsym_update()
no longer needs to treat variables irq_stack_union and __per_cpu_*
specially.

Orabug: 27214992

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>