Purge duplicate symbol names
It is possible for two symbols in a module to have the same name.  For
example, each name might have file scope and come from a different file.
In the case of duplicates, an unambiguous name-to-symbol lookup is not
possible, and so we omit duplicates.

The way this was done was incorrect.  If we read a symbol from
/proc/kallmodsyms that duplicates an existing symbol, we would discard
the new symbol.  There were two problems.  First, the earlier symbol
would still be retained;  name-to-symbol lookup would return that symbol,
which (at least statistically) would probably be the wrong thing to do.
Second, it would mean that address-to-symbol lookup for the discarded
symbol would unnecessarily fail.

The fix is to read symbols in initially.  Once all symbols have been
read, we review the name-to-symbol tables (hash buckets), eliminating all
instances of duplicate names.  The address-to-symbol list is unaffected.

Orabug: 27214992

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>