rtld_db: ensure that an executable has a valid object name
The rtld adds the executable in a process with l_name being the empty
string, which the rtld_db support code in libproc does not handle
correctly.  Regardless, an empty string is causing matching of probe
specifications to fail as well.

This commit ensures that DTrace userspace has a valid l_name for the
executable so that the probe specification matching (on module name)
works as designed.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>