runtest: handle interruptions better
When we are SIGINTed (or HUPped or SIGQUIT), we want to notify the user
that we have done so, in addition to closing things down.  We do *not*
want to remove the $tmpdir: the user might well want to see it in this
situation.

The previous closing-down was also buggy: many kills, in particular the
one in some versions of bash, will respond to one process failing to be
signalled by not attempting to signal any of the others.  We want to try
to kill everything no matter what, so issue the kills one by one in a
for loop.  If we end up invoking this code more than once (as sometimes
happens for INT versus EXIT traps or INT traps received by multiple
subshells), only issue a kill for one of them: issuing the kill more
than once is at best a waste of time and at worst (when pids have
recycled) can be dangerous.

Also, migrate a bunch of shared code into a function, and drop the
SIGSEGV traps: if bash dumps core we cannot trust it to do anything at
all.  (This does not affect the handling of situations when dtrace
itself dumps core.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Orabug: 27577593