Comment out unported parts of dt_pid.
The upcoming changes to dt_proc.c and the libproc proxies require *all*
use of libproc functions to go through dt_proc proxies: rather than
introduce dozens of proxies to support a pid provider which is
nonfunctional anyway, just dike most of it out of the build and
reincorporate it when the pid provider is ported and we know what
proxies are needed.

Enough of dt_pid is retained to allow USDT to continue working.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>