Revamp dt_proc public interfaces to avoid exec() race
The ps_prochandle resetting code added in commit d930a83301 was fatally
flawed.  It tried to arrange for ps_prochandles to be reseatable by
passing them around in a structure containing a ps_prochandle, so that
when an exec() happened while deep in a P*() function, we could reseat
the pointer and keep on going.  This worked, but unfortunately the
DTrace process-control thread can also detect exec() and reseat the
ps_prochandle on its own (via the landing pad in dt_proc_control()), and
this can happen after a reseatable dtrace_prochandle has been acquired
by the master thread but before it does anything with it:

Main thread                Process-control thread
P = dt_proc_grab();
                           exec! dt_proc_reattach(), freeing P.P
dt_Pdosomething(P) -> accesses P.P

Accessing freed memory in a program running as root is a bad idea, but
even if it weren't, even if we will usually be lucky and the memory
won't have been recycled by a later allocation, we will be inspecting
and modifying the address space of some exec()ed program as if it hadn't
exec()ed.  This is unlikely to end in sunshine and rainbows.

Any interface using pointers in the way this does is similarly
vulnerable to this kind of race, so we must discard the
dtrace_prochandle as a bad idea and replace it with a scheme where we
pass down PIDs and look them up in the dph every time they are needed.
This is unlikely to be ruinous performance-wise compared to even a single
ptrace() call or cross-thread proxying operation, and is safe against
pid reuse because outside grabbing (which of course already used pids),
the only pids we can call this with are ones that DTrace is monitoring,
which means they cannot have died yet: if they have died, they will be
detected as dead even if their PID has been recycled.  To allow for
future expansion, we wrap the PIDs in an opaque structure and add an
accessor to let users get the PID out of it.  (This means that users
have to remember to release processes they have grabbed, or leak memory.)

Unfortunately this means an incompatible API/ABI change to libdtrace:
all consumers need source changes and relinking.  We change the names of
dtrace_proc_{create,grab}() to force this (to
dtrace_proc_{create,grab}_pid()): old consumers will fail immediately
when invoked.  (This name is not ideal, since dtrace_probe_create_pid()
has no PIDs anywhere in its interface, but we must change it and the
name should be consistent with dtrace_proc_grab_pid().)

In the world outside dt_proc.c, we make the following changes:

 - All dt_proc functions that took a dtrace_prochandle *now take a pid_t
   instead.  All functions that returned one now return a pid_t instead:
   on error, a value < 0, rather than NULL.  This includes all the
   dt_P*() proxying functions.  ps_prochandles are no longer accessible
   via dt_proc interfaces unless you go grubbing around in the
   dt_proc_t: it is not safe to use them directly.  The
   dtrace_prochandle also takes a PID now.
 - All the dtrace_proc_*() functions now take and return a
   dtrace_proc *, which is opaque: there is a new function
   dtrace_proc_getpid() to get the PID back out.
 - dt_proc_lookup() has lost its last argument, which was only valid
   internally to dt_proc_destroy() in any case.
 - All the dt_ps_*() functions, both inside and outside dt_proc.c, have
   lost their ps_ part: so dt_ps_proc_continue() has become
   dt_proc_continue().  Most of these are unchanged, but functions like
   dt_ps_proc_continue() which were part of the public interface of
   dt_proc have shifted to taking a dt_proc_t instead.
 - dt_proc_create() is no longer public: it had no users outside dt_proc.
   dtrace_proc_create() is still public, but of course returns a
   dtrace_proc * now.
 - dt_proc_grab() and dt_proc_release() are replaced with new helpers
   dt_proc_grab_lock() and dt_proc_release_lock(), which act like
   a dt_proc_grab()/dt_proc_lock() or dt_proc_unlock()/dt_proc_release()
   pair.  (This does not help with raciness, but does make the code a bit
   clearer, since you have to lock right after grabbing and before
   releasing in any case.)

dt_pid.c has not been fully adjusted to the new world: only the USDT
portions have been ported.

Internally to dt_proc.c, there is a lot of churn, but most of it is just
renaming to drop dt_ps_*() names and drop wrappers that used a
dtrace_prochandle or change them to pass in a pid instead.  Substantive
changes outside that:

 - dt_proc_dpr_{lock,unlock}() have lost the redundant _dpr in their
   names.
 - dt_proc_lookup() has been turned into a wrapper that calls
   dt_proc_lookup_remove(), which takes the third 'remove' argument that
   dt_proc_lookup() used to.  The only caller of dt_proc_lookup_remove()
   is dt_proc_destroy(), so it is easy to prove that nothing is doing a
   removal operation that isn't supposed to.  Both functions now look up
   a pid and return a dt_proc_t, rather than looking up a ps_prochandle:
   nearly all callers already had a pid or a dt_proc_t in any case.  An
   assertion that the dpr_proc equalled the ps_prochandle we were
   passing in is now unnecessary (the corresponding assertion, that the
   PID recorded equalled the one we were passing in, is proved by the
   fact that we just looked it up using it).
 - Several functions that used to take a ps_prochandle only to look up a
   dt_proc_t on it, or that used to return a ps_prochandle similarly,
   have been revamped to just take or return a dt_proc_t:
   dt_proc_create(), dt_proc_destroy(), dt_proc_grab(), dt_proc_release(),
   dt_proc_continue().
 - The old dtrace_prochandle-taking wrappers for dt_proc_create(),
   dt_proc_grab(), dt_proc_release(), dt_proc_continue(), dt_proc_lock(),
   dt_proc_unlock() are removed (and mostly replaced with versions that
   take PIDs or dtrace_proc *, depending on whether they are in the
   public interface or the interface internal to libdtrace).
 - All the proxies have been rejigged to take a PID instead of a
   dtrace_prochandle.  You can now define a proxy with zero arguments
   (and we now define one for Pstate()): this means we need to turn off
   -Wpedantic for this file to disable the C99 warning about this GCC
   extension.  GCC 4.4 does not support disabling this warning, so there
   is an extra useless warning on OL6.  Sorry.

Orabug: 27501199
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>