Bump soname.
This lets us rename the misnamed dtrace_proc_create_pid() back to
dtrace_proc_create() (it doesn't have anything to do with pids).
dtrace_proc_grab_pid() retains its new name to leave room for other
sorts of grabs (cross-VM, etc) in the future.

This means we now have a 'non-development' (non-0) soname, but this is
not actually a problem: we've been operating under stable soname rules
for many years, breaking A[BP]I only when we have to.  We've certainly
broken both a lot less than Poppler, which is at something like .so.75
by this point.  We will not run out of numbers, and as long as we don't
make a habit of breaking API or ABI at every release, I think we can
move to .so.1 without worry (and without the ridiculous contortions
I had to use in recent patches to try to get programs using newly-
incompatible APIs to fail to run).

The DTrace consumer API is not glibc and does not have to be quite as
stable as glibc.  Stability is still desirable, of course: we should
certainly start bumping sonames when we do so to try to help people
know when we've had a compatibility break.

Orabug: 27501199
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>