runtest: automatic reinvocation of specific erratic tests
This change adds a new test option, @@reinvoke-failure, which
if present and nonzero is the number of times to reinvoke a
failing test before concluding that it has failed.  Reinvoked
tests which eventually pass are reported as such, and all runs
are logged in the logfile, but are considered passes: the sumfile
contains only the final, succeessful invocation.

This is intended for tests like those in test/unittest/aggs where
DTrace limitations cause the tests to produce subtly different
output on a minority of runs, but where we would like to be sure
that the results are as expected nearly all the time.  It shouldn't
have much impact on speed, because reinvocation only happens if a
test fails, and most tests do not fail.

We turn it on for all tests in test/demo/intro and test/unittest/aggs,
those being the directories most afflicted with spurious test failures
with known cause.

Never turn this option on globally, since it will hide real bugs.

(The diff is scary, but this is mostly due to reindentation. The
actual change is minimal, plus a bit of refactoring since we now
have to emit results from two places, once only to the logfile
and once to both the logfile and sumfile.)

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>