libdtrace-ctf: delete submodule
This is much more convenient to deal with as a separate project: we
never remember to update the submodule link when we need to, so it is
almost always broken in one way or another, in addition to containing
a submodule URL that is only valid inside the Oracle network.

Just delete it.