Adjust interpretation of /proc/kallmodsyms for ARM
Two adjustments in interpreting /proc/kallmodsyms are needed on ARM.

The file first lists kernel (and built-in-module) symbols and then
loaded-module symbols.  There is a separator symbol at the end of the
kernel portion of the file;  its address is the end of kernel address
space.  Function dt_modsym_update() used to check for loaded modules
by comparing a symbol address to the separator address.  But on ARM,
loaded modules are lower in address space than the kernel.  So, the
check has been changed to simply whether a symbol appears after the
separator in the file.

/proc/kallmodsyms also offers clues whether symbols sizes are
correctly reported.  Older kernels had incorrect information
and so certain tests are allowed to XFAIL.  Only newer, patched
kernels are required to pass.  The check on /proc/kallmodsyms
assumed a particular separator symbol name.  It is more robust,
when running on multiple processors (x86, SPARC, ARM), to check
for multiple names (_end and __brk_limit).

Orabug: 27214992

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>