tests: dike out tests that modprobe sdt and dt_test on arm for now
Some tests unload and reload sdt and dt_test as part of their normal
operation.  This works fine unless these modules were not in
test/modules for that arch at all, in which case we end up with a module
loaded that was never loaded at testsuite startup, and is not unloaded
at the end.  If this module uses dtrace.ko (as sdt and dt_test do), the
rmmod of dtrace.ko then fails with an ugly message.

Skip those tests on such arches.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>