dtrace: clean up gcc compiler warnings on x86
Building dtrace-utils would typically lead to a variety of gcc compiler
warnings.  At the very least, this introduces noise into the build
process.  This commit fixes many of those warnings on x86.  The most
common case is
    warning: <variable> may be used uninitialized in this function
    [-Wmaybe-uninitialized]

One case that remains is the use of constructor priority 0 in
include/sys/compiler.h.  This priority is needed, and the issue
has been documented in that file.

Orabug: 27934422

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>