Support for USDT and pid probes on aarch64
This commit adds support for USDT and pid provider probes for arm64.
Due to the fixed instruction size and the way uprobes handles the
out-of-line execution (or emulation) of the instruction that a probe
is placed on, arm64 needs the same mechanism as is used for sparc64.
Specifically, this means that we pass a 0 argument in the (fake) call
for is-enabled probes.  This uses the fact that the same register is
used for the first argument to calls and the return value from calls.
Passing in 0 ensures that the default value in the return value
register will be 0 (false condition for the is-enabled probe).

Both regular probes and is-enabled probes can occur as regular calls
and tail calls.  The first case shows us as R_AARCH64_CALL26 reloc
records, and the latter as R_AARCH64_JUMP26 reloc records.  A regular
call can be changed into a NOP, whereas a tail call needs to become
a RET (because it needs to preserve the semantics of a tail call).

Various tests that are meant to exercise the pid provider depend on
SDT probes working, which is not the case (yet) on arm64.  There are
being marked as XFAIL for now.

This patch also introduces support in showUSDT to display relocation
records.

Orabug: 27847969
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>