Make tst.dlclose1.sh more robust
The tst.dlclose1.sh test depends on the implementation of the pause
function and the underlying system call.  Architectures currently have
one of two possible implementations, one based on a pause syscall, or
one based on rt_sig* syscalls.  The test has been modified to support
both rather than relying on arch-dependent conditionals.  This also
ensures that when an architecture implementation moves from one to the
other, the test keeps working.

Orabug: 27847969
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>