err.D_PROC_CREATEFAIL.many.d causes failures in subsequent tests
The err.D_PROC_CREATEFAIL.many.d test reports hitting the limit of
pid probes that can exist at one time as an out-of-memory condition.
While that is correct based on the DTrace implementation at the kernel
level (albeit perhaps not the best error code), the delayed cleanup
code that frees up slots for pid probes runs interference with the
subsequent tests that depend on being able to create pid probes.

This test needs some form of delay after running so allow the system
to clear out the probes that are pending cleanup, so that further
tests are not impacted.

Orabug: 27847969
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>