dtrace: remove preallocate from buffering testsuite
The buffering stress tests are meant to exercise kernel buffer
size downscaling when not enough memory is present.  On larger
machines the usage of the 'preallocate' option creates huge
stress on the kernel vm that does not allow test to get to the
point where it actually tests the kernel buffer allocation.

At the same time the preallocate option does not have anything
to do with userspace buffer allocation so should be removed from
the kernel buffer tests.

Orabug: 27998779

Signed-off-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>