Do not remove items from the dph before their control thread is dead.
We were handling process destruction by removing items from the dph,
then from the notifier chain, then killing their control thread,
then freeing everything.

This order is wrong, because as of 5cf06de437 proxiable dt_P*()
operations, even when called from the same thread, now have to do a
dt_proc_lookup() to map from the pid to the dpr (where before they had a
dpr right there waiting for them).  This means that they require the dpr
to be available in the dph, so we had better not remove it from the dph
before the process-control thread is stopped.  (After it is stopped we
can remove it safely, because we know no more proxiable operations will
be carried out on it.)

Orabug: 27961105
Fixes: 5cf06de437 ('Revamp dt_proc public interfaces to avoid exec() race')
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>