tests: faster aggmod test
There have been problems converting addresses into symbolic names,
including module names.  The test tst.aggmod_full.sh was introduced
to check that @[mod(addr)] produced exactly one name for each module.
It had one D action for each of the roughly 100,000 symbols in
/proc/kallmodsyms.  Since DTrace has hard-coded limits on script
sizes, /proc/kallmodsyms had to be split into hundreds of pieces,
compromising testing and increasing test time due to the many
invocations of dtrace_open() that resulted.

Here, a new test is introduced.  It is a C program that reads
/proc/kallmodsyms and feeds (addr,modname) pairs to a static probe.
The C program also writes modname, which is aggregated with Linux
command-line tools (sort | uniq -c).  A simple D script takes
(addr,modname) pairs and aggregates [mod(addr),modname].  The
redundant module-name values are checked for consistency and the
output is checked against the C-Linux output.

The test has a slight nanosleep() throttle to keep the test program
from dropping data.

The new test does not yet work on ARM.  So for the time being, the
coverage requirement of the old test was lowered to allow these new,
but slower, processors to pass.  Once the new test passes on ARM, the
the older test should probably be dropped.

Orabug: 28007146

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>