Purge pthread_cancel().
It is almost impossible for dtrace to terminate reliably as long as it
is relying on pthread_cancel() to terminate the process-control thread:
there are numerous cancellation points in loci where it is not at all
safe to terminate (in the middle of breakpoint manipulation, for
instance). Rather than relying on that, signal quitting via a proxy
message and wait for the process-control thread to respond to it (and
terminate).

For now, we wait forever, because the consequences of terminating
without cleaning up breakpoints properly are too serious for
unceremonious (unacknowledged) aborts of our process-control threads.

Orabug: 27961105
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>