tests: manage temporary files more consistently
Place temporary files in consistent locations such as
    $tmpdir/
    $tmpdir/directory-test.$$.$RANDOM/
rather than in
    /tmp/
    ${TMPDIR:-/tmp}/
    /var/tmp/
(Note that $TMPDIR is not typically defined.)

In test/unittest/scripting/tst.*.sh, name the D files consistently with
a .d extension.

In general, change the aberrations to be more consistent with previous
clean-up attempts and with other tests.

Orabug: 28142056

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>