dtrace: runtime option -x noresolve
DTrace runtime option -x noresolve. This option is to prevent
automatic resolution of userspace symbols. It can be
useful when symbol resolution takes a long time for large
statically-linked binaries.

Applications that contain many symbols can incur an overhead when
DTrace resolves addresses to symbol names. This overhead can be
reduced with the noresolve option.

This feature was initially implemented in Solaris 11.3 per
customer request that uses very large statically-linked binaries.

for ustack action: if noresolve option is set, just print the
<object_file>:<address> format.
for uaddr action: if noresolve option is set, just print the
<object_file>:<address> format.
for usym action: if noresolve option is set, just print the
<object_file>:<address> format.

Orabug: 28475467

Signed-off-by: Vincent Lim <vincent.lim@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>