Adjust SDT args test to increased number of possible arguments.
The DTrace code in the kernel is being updated to support up to 18
arguments for SDT probes, so the test to ensure we can retrieve the
SDT probe arguments has been updated as well to the new maximum
number of arguments.

For consistency, we also increase the number of supported arguments
for USDT probes to 18 (so the sdt_internal.h macros remain identical
between kernel and userspace).

Orabug: 28217266
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>