libproc: resume processes properly on thread creation
We currently spot thread creation via PTRACE_O_TRACECLONE so that we can
deactivate things like rtld_db, which are not ready to handle multiple
threads yet.

However, we then forget to resume the main thread (though we do resume
the newly-created thread), so everything deadlocks; and we fail to
resume the child thread correctly, since it will enter PTRACE_EVENT_STOP
state and we have to restart it when it does so, just as for forked
children.  Resume the main thread, and reuse the forked-children
machinery to resume the new thread, so that both threads restart
reliably, and the new thread is detached from once it is safe to do so.

Orabug: 28210986
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>