tests: add some tests for multithreaded processes
These tests verify that tracing an exec()ed multithreaded process does
not hang, and that tracing an exec()ed multithreaded process that
exec()s in the non-main thread does not hang.  This which does not
happen even before the recent fix, but that is sensitive to changes in
the behaviour of waitpid() around new threads.  This has changed before,
so we should have a test ready for its happening again.

One of these tests requires OMP support in the compiler: all versions of
GCC we have ever supported implement OMP.

Orabug: 28210986
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>