pid: zero-initialize the struct fasttrap_probe_spec
Otherwise we leave some members uninitialized.  The kernel happens not
to access those members, but this is a risky thing to rely upon, and
causes valgrind false-positives unless we teach valgrind an unreasonable
amount about the access patterns of this one ioctl().

It is easier just to zero-initialize it consistently.

Orabug: 28247636
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>