dtrace -l: do not pass garbage into the kernel
When called with a NULL pdp, as dtrace -l does, this function calls into
the kernel with an uninitialized struct dtrace_probedesc_t.  It so
happens that the implementation of DTRACE_PROBES inside the kernel makes
no use of the uninitialized fields, but it is intertwined with the
DTRACE_PROBEMATCH implementation, which does make use of them, and we do
*print* them in a debugging statement, and even though we null-terminate
them (at max length) before then this is quite disconcerting and causes
valgrind to (rightly) complain, in addition to likely printing out
garbage if people turn on the relevant bit of debug output.

So unconditionally zero-initialize the dtrace_probedesc_t before
passing it in.

Orabug: 28247636
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Tomas Jedlicka <tomas.jedlicka@oracle.com>