rtld_db: do not crash on self-grabs and on races with termination
A self-grab is a subset of a noninvasive trace, as is used for critical
system daemons: the process (ourself) is grabbed noninvasively, is not
ptraced, and from rtld_db's perspective is essentially "dead" because
it needs ptrace() to do anything useful.

Unfortunately, dt_proc_rdagent() is not really expecting this, and calls
Prd_agent() assuming that it doesn't return NULL in subsequent calls if
it ever returned non-NULL once.  If the grab is still setting up (as is
likely) we may well lose our process-control thread to noninvasiveness
and find that Prd_agent() has flipped from returning non-NULL to
returning NULL.  (This can also happen if a normal non-self-grabbed
process dies at the wrong instant, but this is a very narrow race and
nobody has hit it.)

Even if dt_proc_rdagent() was allowing for this, it would still be
dangerous: process termination frees the rd_agent pointer, and can race
with this call.  So move things around so that rd_agent freeing happens
only when the rest of the process freeing happens, at DTrace shutdown
time, eliminating the race.  Also ensure that various bits of the
ELF-management code detect PS_DEAD or null-pointer processes
appropriately (these are bits used early in the rd_event_enable() call
path, to try to get rd_event_enable() to properly return NULL rather
than coredumping if it *is* called with a null pointer for a rd_agent or
a proc.)

Also added a test for this.  The test fails at compilation time, but
does not coredump, and that is all we're interested in for this bug.

Orabug: 28361373
Signed-off-by: Nick Alcock <nick.alcock@oracle.com>