libproc: properly handle breakpoints which come in while DTrace is exiting
Commit ea1966e9df3ba endeavoured to prevent breakpoints from firing
while they are being torn down, to prevent the handlers from doing
things like adding more breakpoints, which might land in places in the
hash chain that the flusher has already traversed, leading to stale
breakpoints that are never torn down and outstanding deadly traps in the
victim process.

Unfortunately, that's not what it did. The bkpt_consume mechanism
consumes SIGTRAPs, but does not truly handle the resulting breakpoints:
it just ignores the handlers.  If the process hits a breakpoint, it'll
end up stopped on it but without P->tracing_bkpt set: so Punbkpt() will
not realise that it is stopped on a breakpoint, and will not move the
instruction pointer back when resuming the process.  Result: at best, a
missed instruction: more likely, a process resumed mid-instruction, and
a bizarre unreproducible segfault.

Instead of messing about with bkpt_consume, simply use Ptrace() to halt
execution of the process across flushes. A halted process cannot
possibly hit a breakpoint while we're doing the removal, and if we have
hit the breakpoint *already*, bkpt_flush() calls Pwait() later on to
consume it (at a point when the breakpoint handler is gone, so it'll
simply resume the process, but that doesn't matter at all given that
with fractionally different timing it would have hit after the
breakpoint was removed and never fired the handler at all).

Orabug: 28473826