cg: fix up cg_call() return regs
Many of the semi-mechanically translated opcodes take a destination
register as a last argument.  BPF convention has us leaving the input
regs unchanged and passing the results back in %r0.  The verifier relies
upon this, so we'd better not violate it.

Since this means a lot of places are now doing ret-and-mov, add a
dt_cg_mov() to do the gruntwork.  (Note: dt_cg_mov() does not allocate a
reg for you, only modifies an already-allocated one.)