cg: allow immediate args to helpers
We introduce an array of prototypes that describe the arguments accepted
by every BPF helper (in the same order they are declared in
include/uapi/linux/bpf.h).  This lets us pass in regs and immediates to
dt_cg_call(), and generate appropriate moves / immediate loads
automatically.

There is even less checking that we call with the right number of args
than there was before.