bpf: simplify assignment cg; implement the dot operator
BPF can add immediates to registers without being constrained by DIF's
ridiculous 16-bit immediate limitation, so we can discard an unnecessary
temporary in the implementation of assignments of structure members and
just add the structure member offset directly in one instruction.

While we're at it, implement the dot operator the same way (which was
put off until translators were done, but turned out not to depend on the
details of translators at all).